Backend/requirements.txt
Flask>=1.0.0
cloudant==2.13.0
Flask-Cors==3.0.10
ibm-cos-sdk==2.3.2
gunicorn==20.1.0
Flask-RESTful==0.3.9
flask-restful==0.3.9
flask-restplus==0.13.0
python-dotenv == 0.19.0
jsonschema == 3.2.0
flask-expects-json == 1.7.0
pyjwt == 2.3.0
bcrypt == 3.2.0
decorate-all-methods == 0.0.2
flasgger == 0.9.1